Fix up ENTRYPOINT/CMD to be more flexible.
The ENTRYPOINT is now used for the LICENSE check, at which point the CMD
takes over and is executed directly via exec.  This also enables the
user to run arbitrary commands such as "productInfo version" directly.
It also enables a data volume container to be created directly from the
websphere-liberty image rather than something else based on ubuntu (to
get around the original ENTRYPOINT limitation).