Add common image. Move to August Beta.
Insert a new common image between kernel and the webProfile7 image that adds the base set of production features.

Modify webProfile6 to build using the common image as a starting point. Explicitly pull in additional features to that it matches the contents of the previous runtime JAR.

Move up to the August beta.