Fix potential max heap issues with OpenJ9 SCCs (#344)
* Minor populate_scc.sh cleanups

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>

* Generate SCCs for small heaps

The compressedrefs mode, AOT code saved in the OpenJ9 SCC is tailored
to the heap size and position of the running application. If size and/or
position change the code in the SCC may not be loaded and new code may
not be saved. Recent OpenJDK JDK8 builds use different criteria for
choosing heap size, which may increase the odds of running with
incompatile AOT code. This patch adds an option that reverts the new
criteria for JDK8 when generating the SCCs, thereby producing a more
compatible SCC, on average.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>

* Update README with SCC info

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>