Reduce SCC build time for app images (#378)
Currently every image in the stack runs the server 4 times to build
its SCC layer. The final 2 iterations are optional but allow for the
refinement of profiling information. This patch reduces the number
of optional iterations for app images from 2 to 1 to reduce build
times on the basis that it does not sacrifice much, if any,
start-up performance. It does not affect the number of iterations
used for server SCC layers.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>